+++ src/agent/transform-headers-agent.ts
@@ -1,25 +1,24 @@
 /* eslint-disable no-underscore-dangle */
-const HeaderGenerator = require('header-generator');
-const http = require('http');
-const WrappedAgent = require('./wrapped-agent');
+// @ts-expect-error TODO: Type `header-generator`
+import HeaderGenerator from 'header-generator';
+import { OutgoingMessage, Agent, ClientRequest, ClientRequestArgs } from 'http';
+import { WrappedAgent } from './wrapped-agent';
 
-const { _storeHeader } = http.OutgoingMessage.prototype;
+// @ts-expect-error Private property
+const { _storeHeader } = OutgoingMessage.prototype;
 
 const generator = new HeaderGenerator();
 
 /**
- * @description Transforms the casing of the headers to Pascal-Case.
+ * Transforms the casing of the headers to Pascal-Case.
  */
-class TransformHeadersAgent extends WrappedAgent {
+export class TransformHeadersAgent<T extends Agent> extends WrappedAgent<T> {
     // Rewritten from https://github.com/nodejs/node/blob/533cafcf7e3ab72e98a2478bc69aedfdf06d3a5e/lib/_http_outgoing.js#L442-L479
     /**
-     * @description Transforms the request via header normalization.
-     * @see {TransformHeadersAgent.toPascalCase}
-     * @param {http.ClientRequest} request
-     * @param {boolean} sortHeaders - if the headers should be sorted or not
+     * Transforms the request via header normalization.
      */
-    transformRequest(request, sortHeaders) {
-        const headers = {};
+    transformRequest(request: ClientRequest, { sortHeaders }: {sortHeaders: boolean}): void {
+        const headers: Record<string, string | number | string[]> = {};
         const hasConnection = request.hasHeader('connection');
         const hasContentLength = request.hasHeader('content-length');
         const hasTransferEncoding = request.hasHeader('transfer-encoding');
@@ -28,9 +27,9 @@ class TransformHeadersAgent extends WrappedAgent {
 
         for (const key of keys) {
             if (key.toLowerCase().startsWith('x-')) {
-                headers[key] = request.getHeader(key);
+                headers[key] = request.getHeader(key)!;
             } else {
-                headers[this.toPascalCase(key)] = request.getHeader(key);
+                headers[this.toPascalCase(key)] = request.getHeader(key)!;
             }
 
             if (sortHeaders) {
@@ -39,8 +38,15 @@ class TransformHeadersAgent extends WrappedAgent {
             }
         }
 
+        const typedRequest = request as ClientRequest & {
+            _removedContLen: boolean;
+            _contentLength: number;
+            _removedTE: boolean;
+            agent?: Agent;
+        };
+
         if (!hasConnection) {
-            const shouldSendKeepAlive = request.shouldKeepAlive && (hasContentLength || request.useChunkedEncodingByDefault || request.agent);
+            const shouldSendKeepAlive = request.shouldKeepAlive && (hasContentLength || request.useChunkedEncodingByDefault || typedRequest.agent);
             if (shouldSendKeepAlive) {
                 headers.Connection = 'keep-alive';
             } else {
@@ -54,48 +60,45 @@ class TransformHeadersAgent extends WrappedAgent {
             //
             // Note: This uses private `_removedTE` property.
             //       This property tells us whether the transfer-encoding was explicitly removed or not.
-            if (!hasTrailer && !request._removedContLen && typeof request._contentLength === 'number') {
-                headers['Content-Length'] = request._contentLength;
-            } else if (!request._removedTE) {
+            if (!hasTrailer && !typedRequest._removedContLen && typeof typedRequest._contentLength === 'number') {
+                headers['Content-Length'] = typedRequest._contentLength;
+            } else if (!typedRequest._removedTE) {
                 headers['Transfer-Encoding'] = 'chunked';
             }
         }
 
-        const transformedHeaders = sortHeaders ? generator.orderHeaders(headers) : headers;
+        const transformedHeaders: Record<string, string | number | string[]> = sortHeaders ? generator.orderHeaders(headers) : headers;
 
-        // eslint-disable-next-line no-restricted-syntax, guard-for-in
-        for (const key in transformedHeaders) {
-            request.setHeader(key, transformedHeaders[key]);
+        for (const [key, value] of Object.entries(transformedHeaders)) {
+            request.setHeader(key, value);
         }
     }
 
-    addRequest(request, options) {
+    override addRequest(request: ClientRequest, options: ClientRequestArgs): void {
+        const typedRequest = request as ClientRequest & {
+            _storeHeader: (firstLine: string, headers: Record<string, string>) => void;
+        };
+
         // See https://github.com/nodejs/node/blob/533cafcf7e3ab72e98a2478bc69aedfdf06d3a5e/lib/_http_outgoing.js#L373
         // Note: This overrides the private `_storeHeader`.
         //       This is required, because the function copies
         //       the `connection`, `content-length` and `trasfer-encoding` headers
         //       directly to the underlying buffer.
-        request._storeHeader = (...args) => {
-            this.transformRequest(request, true);
+        typedRequest._storeHeader = (...args) => {
+            this.transformRequest(request, { sortHeaders: true });
 
             return _storeHeader.call(request, ...args);
         };
 
         // `agent-base` isn't able to detect the protocol correctly
-        options.secureEndpoint = options.protocol === 'https:';
+        (options as any).secureEndpoint = options.protocol === 'https:';
 
         return super.addRequest(request, options);
     }
 
-    /**
-     * @param {string} header - header with unknown casing
-     * @returns {string} - header in Pascal-Case
-     */
-    toPascalCase(header) {
+    toPascalCase(header: string): string {
         return header.split('-').map((part) => {
-            return part[0].toUpperCase() + part.slice(1).toLowerCase();
+            return part[0]!.toUpperCase() + part.slice(1).toLowerCase();
         }).join('-');
     }
 }
-
-module.exports = TransformHeadersAgent;