+++ src/hooks/options-validation.ts
@@ -1,9 +1,6 @@
-const { default: ow } = require('ow');
+import ow from 'ow';
 
-/**
- * @param {object} options
- */
-function optionsValidationHandler(options) {
+export function optionsValidationHandler(options: unknown): void {
     const validationSchema = {
         proxyUrl: ow.optional.string.url,
         useHeaderGenerator: ow.optional.boolean,
@@ -13,7 +10,3 @@ function optionsValidationHandler(options) {
 
     ow(options, ow.object.partialShape(validationSchema));
 }
-
-module.exports = {
-    optionsValidationHandler,
-};