+++ src/scraping-defaults.ts
@@ -1,6 +1,6 @@
-const crypto = require('crypto');
+import { constants } from 'crypto';
 
-const SCRAPING_DEFAULT_OPTIONS = {
+export const SCRAPING_DEFAULT_OPTIONS = {
     // Most of the new browsers use HTTP2
     http2: true,
     https: {
@@ -23,17 +23,12 @@ const SCRAPING_DEFAULT_OPTIONS = {
 
 /**
  * Reorders the default NodeJs ciphers so the request tries to negotiate the modern TLS version first, same as browsers do.
- * @returns {string} ciphers list
  */
 function ensureModernTlsFirst() {
     const modernTlsCiphers = ['TLS_AES_256_GCM_SHA384', 'TLS_AES_128_GCM_SHA256', 'TLS_CHACHA20_POLY1305_SHA256'];
-    const defaultCiphers = new Set(crypto.constants.defaultCipherList.split(':'));
+    const defaultCiphers = new Set(constants.defaultCipherList.split(':'));
     // First we will remove the modern ciphers from the set.
     modernTlsCiphers.forEach((cipher) => defaultCiphers.delete(cipher));
     // Then we will add the modern ciphers at the beginning
     return modernTlsCiphers.concat(Array.from(defaultCiphers)).join(':');
 }
-
-module.exports = {
-    SCRAPING_DEFAULT_OPTIONS,
-};