+++ test/agent.test.ts
@@ -1,20 +1,20 @@
-const http = require('http');
-const getStream = require('get-stream');
-const TransformHeadersAgent = require('../src/agent/transform-headers-agent');
+import { AddressInfo } from 'net';
+import http, { Server } from 'http';
+import getStream from 'get-stream';
+import { TransformHeadersAgent } from '../src/agent/transform-headers-agent';
+import { startDummyServer } from './helpers/dummy-server';
 
 const agent = new http.Agent({
     keepAlive: true,
 });
 
-const { startDummyServer } = require('./helpers/dummy-server');
-
 describe('TransformHeadersAgent', () => {
-    let server;
-    let port;
+    let server: Server;
+    let port: number;
 
     beforeAll(async () => {
         server = await startDummyServer();
-        port = server.address().port;
+        port = (server.address() as AddressInfo).port;
     });
 
     afterAll(() => {
@@ -48,7 +48,7 @@ describe('TransformHeadersAgent', () => {
         });
 
         const transformAgent = new TransformHeadersAgent(agent);
-        transformAgent.transformRequest(request);
+        transformAgent.transformRequest(request, { sortHeaders: true });
 
         request.end();
     });
@@ -70,7 +70,7 @@ describe('TransformHeadersAgent', () => {
         });
 
         const transformAgent = new TransformHeadersAgent(agent);
-        transformAgent.transformRequest(request);
+        transformAgent.transformRequest(request, { sortHeaders: true });
 
         request.end();
     });
@@ -103,7 +103,6 @@ describe('TransformHeadersAgent', () => {
 
         const request = http.request(`http://localhost:${port}/headers`, {
             agent: transformAgent,
-            sortedHeaders: ['Host'],
         }, async (response) => {
             const body = await getStream(response);
             const headers = JSON.parse(body);