+++ CHANGELOG.md
@@ -1,7 +1,18 @@
+3.0.0 / TBD
+====================
+- Switch to TypeScript
+- Enable insecure parser by default
+- Use `header-generator` to order headers
+- Migrate to `hpagent`
+- Remove `default` export in favor of `import { gotScraping }`
+
+2.1.2 / 2021/08/06
+====================
+- Mimic `got` interface
+
 2.1.1 / 2021/08/06
 ====================
 - Use `header-generator` v1.0.0
-
 
 2.1.0 / 2021/08/06
 ====================